Quang

+++ src/services/PostService.ts
@@ -5,13 +5,16 @@ import StatusCodeEnum from "../enums/StatusCodeEnum";
 import Database from "../utils/database";
 import { IQuery } from "../interfaces/IQuery";
 import { cleanUpOldAttachments } from "../utils/filePathFormater";
-
+import UserRepository from "../repositories/UserRepository";
+import UserEnum from "../enums/UserEnum";
 class PostService {
   private postRepository: PostRepository;
   private database: Database;
+  private userRepository: UserRepository;
   constructor() {
     this.postRepository = new PostRepository();
     this.database = Database.getInstance();
+    this.userRepository = new UserRepository();
   }
   createPost = async (
     userId: string | ObjectId,
@@ -20,6 +23,7 @@ class PostService {
     attachments: Array<string>
   ) => {
     const session = await this.database.startTransaction();
+
     try {
       const post = await this.postRepository.createPost(
         {
@@ -30,6 +34,7 @@ class PostService {
         },
         session
       );
+
       await this.database.commitTransaction();
       return post;
     } catch (error) {
@@ -43,9 +48,26 @@ class PostService {
       );
     }
   };
-  getPost = async (id: string | ObjectId) => {
+
+  getPost = async (id: string | ObjectId, requesterId: string) => {
     try {
-      const post = await this.postRepository.getPost(id);
+      let ignoreDeleted = false;
+      const checkRequester = await this.userRepository.getUserById(
+        requesterId,
+        ignoreDeleted
+      );
+      if (!checkRequester) {
+        throw new CustomException(
+          StatusCodeEnum.NotFound_404,
+          "Requester not found"
+        );
+      }
+      if (
+        [UserEnum.ADMIN, UserEnum.SUPER_ADMIN].includes(checkRequester?.role)
+      ) {
+        ignoreDeleted = true;
+      }
+      const post = await this.postRepository.getPost(id, ignoreDeleted);
       return post;
     } catch (error) {
       if (error as Error | CustomException) {
@@ -58,9 +80,25 @@ class PostService {
     }
   };
 
-  getPosts = async (query: IQuery) => {
+  getPosts = async (query: IQuery, requesterId: string) => {
     try {
-      const posts = await this.postRepository.getPosts(query);
+      let ignoreDeleted = false;
+      const checkRequester = await this.userRepository.getUserById(
+        requesterId,
+        ignoreDeleted
+      );
+      if (!checkRequester) {
+        throw new CustomException(
+          StatusCodeEnum.NotFound_404,
+          "Requester not found"
+        );
+      }
+      if (
+        [UserEnum.ADMIN, UserEnum.SUPER_ADMIN].includes(checkRequester?.role)
+      ) {
+        ignoreDeleted = true;
+      }
+      const posts = await this.postRepository.getPosts(query, ignoreDeleted);
       return posts;
     } catch (error) {
       if (error as Error | CustomException) {
@@ -77,11 +115,18 @@ class PostService {
     id: string | ObjectId,
     title: string,
     content: string,
-    attachments: Array<string>
+    attachments: Array<string>,
+    requesterId: string
   ) => {
     const session = await this.database.startTransaction();
     try {
-      const oldPost = await this.postRepository.getPost(id);
+      const oldPost = await this.postRepository.getPost(id, true);
+      if (requesterId.toString() !== oldPost.userId.toString()) {
+        throw new CustomException(
+          StatusCodeEnum.Forbidden_403,
+          "You are not allowed to update this post"
+        );
+      }
       const post = await this.postRepository.updatePost(
         id,
         {
@@ -91,8 +136,11 @@ class PostService {
         },
         session
       );
+
       await this.database.commitTransaction();
+
       cleanUpOldAttachments(oldPost.attachments);
+
       return post;
     } catch (error) {
       await this.database.abortTransaction();
@@ -105,9 +153,18 @@ class PostService {
       );
     }
   };
-  deletePost = async (id: string | ObjectId) => {
+
+  deletePost = async (id: string | ObjectId, requesterId: string) => {
     const session = await this.database.startTransaction();
+
     try {
+      const oldPost = await this.postRepository.getPost(id, true);
+      if (requesterId.toString() !== oldPost.userId.toString()) {
+        throw new CustomException(
+          StatusCodeEnum.Forbidden_403,
+          "You are not allowed to update this post"
+        );
+      }
       const post = await this.postRepository.deletePost(id, session);
       await this.database.commitTransaction();
       return post;
